+++ Texture.h
@@ -23,7 +23,7 @@ private:
 	float height;
 public:
 	Texture();
-	Texture(unsigned int id, TextureType type, const std::string& path, float width = 0.0f, float height = 0.0f, int channelsCount = 3);
+	Texture(unsigned int id, TextureType type, const std::string& path, int width = 0.0f, int height = 0.0f, int channelsCount = 3);
 	unsigned int GetId();
 	TextureType GetType();
 	std::string GetPath();
@@ -32,9 +32,9 @@ public:
 	float GetHeight();
 	void SetType(TextureType type);
 	void SetPath(const std::string& path);
-	void SetSize(float width, float height);
-	void SetWidth(float width);
-	void SetHeight(float height);
+	void SetSize(int width, int height);
+	void SetWidth(int width);
+	void SetHeight(int height);
 	static Texture LoadTexture(const std::string& path, bool useSRGB = false);
 	static Texture CreateEmptyTexture(int width, int height, TextureType type = TextureType::COLOR);
 	static unsigned int LoadCubeMap(const std::vector<std::string>& pathes);
